mara_api/requirements.txt
fastapi[all]
uvicorn
google-generativeai
python-jose[cryptography]
passlib[bcrypt]
celery
redis